# docs/swagger.yaml
definitions:
...
  models.AccountType:
    properties:
      _id:
        type: string
      name:
        type: string
    type: object
...
  models.Currency:
    properties:
      _id:
        type: string
      iso:
        type: string
      name:
        type: string
      symbol:
        type: string
    type: object
  models.Employee:
    properties:
      _id:
        type: integer
      fullName:
        type: string
      role:
        type: integer
    type: object
  models.EmployeePipeline:
    properties:
      price:
        type: number
      quantity:
        type: integer
      total:
        type: number
      uid:
        $ref: '#/definitions/models.Employee'
      valid:
        type: integer
    type: object
...
  models.Location:
    properties:
      _id:
        type: string
      iso:
        type: string
      name:
        type: string
    type: object
...
  models.TableData:
    properties:
      _id:
        type: string
      baseCurrency:
        $ref: '#/definitions/models.Currency'
      basePrice:
        type: number
      baseRate:
        type: number
      baseTotal:
        type: number
      buyer:
        $ref: '#/definitions/models.Employee'
      cancellationCause:
        type: string
      cancelledBy:
        $ref: '#/definitions/models.Employee'
      completedBy:
        $ref: '#/definitions/models.Employee'
      crossRate:
        type: number
      currency:
        $ref: '#/definitions/models.Currency'
      dateCancelled:
        type: integer
      dateCompleted:
        type: integer
      dateCreated:
        type: integer
      dateReturned:
        type: integer
      dateTaken:
        type: integer
      dateUpdated:
        type: integer
      description:
        type: string
      farmer:
        $ref: '#/definitions/models.Employee'
      fileName:
        type: string
      location:
        $ref: '#/definitions/models.Location'
      price:
        type: number
      quantity:
        type: integer
      rate:
        type: number
      returnedBy:
        $ref: '#/definitions/models.Employee'
      status:
        type: integer
      takenBy:
        $ref: '#/definitions/models.Employee'
      team:
        $ref: '#/definitions/models.Team'
      total:
        type: number
      type:
        $ref: '#/definitions/models.AccountType'
      updatedBy:
        $ref: '#/definitions/models.Employee'
      valid:
        type: integer
    type: object
...
  models.Team:
    properties:
      id:
        type: integer
      teamlead:
        $ref: '#/definitions/models.Employee'
    type: object
...
info:
  contact: {}
  description: Farming service API description.
  title: Gipsyland Farming
  version: "2.0"
paths:
...
  /v2/locations/create:
    post:
      consumes:
      - application/json
      description: creates location
      parameters:
      - description: location body json
        in: body
        name: id
        required: true
        schema:
          $ref: '#/definitions/models.Location'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Create location
      tags:
      - Locations
  /v2/locations/delete:
    delete:
      consumes:
      - application/json
      description: Deletes location by id
      parameters:
      - description: location id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Delete location
      tags:
      - Locations
...
  /v2/locations/get:
    get:
      consumes:
      - application/json
      description: get location by id
      parameters:
      - description: location id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Location'
      summary: Get location
      tags:
      - Locations
...
  /v2/locations/update:
    patch:
      consumes:
      - application/json
      description: Updates location by id
      parameters:
      - description: location body
        in: body
        name: id
        required: true
        schema:
          $ref: '#/definitions/models.Location'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Update location
      tags:
      - Locations
  /v2/tableData/aggregate/buyers:
    get:
      consumes:
      - application/json
      description: returns buyer list
      parameters:
      - description: period start date
        in: query
        name: startDate
        type: string
      - description: period end date
        in: query
        name: endDate
        type: string
      - description: teamlead id
        in: query
        name: teamleadID
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.EmployeePipeline'
            type: array
      summary: Get buyer list
      tags:
      - Table data
  /v2/tableData/aggregate/farmers:
    get:
      consumes:
      - application/json
      description: returns farmer list
      parameters:
      - description: period start date
        in: query
        name: startDate
        type: string
      - description: period end date
        in: query
        name: endDate
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.EmployeePipeline'
            type: array
      summary: Get farmer list
      tags:
      - Table data
  /v2/tableData/aggregate/teamleads:
    get:
      consumes:
      - application/json
      description: returns team list
      parameters:
      - description: period start date
        in: query
        name: startDate
        type: string
      - description: period end date
        in: query
        name: endDate
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.EmployeePipeline'
            type: array
      summary: Get team list
      tags:
      - Table data
  /v2/tableData/get:
    get:
      consumes:
      - application/json
      description: returns all account requests by period and employee
      parameters:
      - description: period start date
        in: query
        name: startDate
        type: string
      - description: period end date
        in: query
        name: endDate
        type: string
      - description: status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.TableData'
            type: array
      summary: Get account requests by period and employee
      tags:
      - Table data
...
swagger: "2.0"